+++ src/Ext/House/Body.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <HouseClass.h>
+#include <TActionClass.h >
 
 #include <Helpers/Macro.h>
 #include <Utilities/Container.h>
@@ -51,4 +52,5 @@ public:
 	static int ActiveHarvesterCount(HouseClass* pThis);
 	static int TotalHarvesterCount(HouseClass* pThis);
 	static HouseClass* GetHouseKind(OwnerHouseKind kind, bool allowRandom, HouseClass* pDefault, HouseClass* pInvoker = nullptr, HouseClass* pVictim = nullptr);
+	static int GetHouseIndex(int param, TeamClass* pTeam, TActionClass* pTAction);
 };